requirements.txt
flask
flask-cors
python-dotenv
tensorflow
faiss-cpu
requests
numpy
opencv-python
